3/requirements.txt
matplotlib>=3.0.3
Pillow>=6.1.0
tensorflow>=1.14.0
setuptools>=41.0.0
opencv-python==3.3.0.10
opencv-contrib-python==3.3.0.10
keras>=2.2.4
